administration page

+++ testProject/src/main/webapp/jsp/admin.jsp
@@ -20,31 +20,45 @@ add users in table and some additional information about users actions<br>
 <table class="sortable" width="100px;">
     <thead>
     <tr>
-        <th>id</th>
         <th>User name</th>
         <th>Password</th>
+        <th>Home directory</th>
         <th>Enabled</th>
+        <th>Write</th>
+        <th>Idle time</th>
+        <th>Upload rate</th>
+        <th>Download rate</th>
+        <th>Max login number</th>
+        <th>Max login IP</th>
+        <th>Email</th>
         <th>User role</th>
         <th>Action</th>
     </tr>
     </thead>
     <s:iterator value="userForm.usersList" var="user">
         <s:url var="userEditURL" action="editUser">
-            <s:param name="userForm.id">
-                <s:property value="#user.id"/>
+            <s:param name="userForm.user_id">
+                <s:property value="#user.user_id"/>
             </s:param>
         </s:url>
         <s:url var="userDeleteURL" action="deleteUser">
-            <s:param name="userForm.id">
-                <s:property value="#user.id"/>
+            <s:param name="userForm.user_id">
+                <s:property value="#user.user_id"/>
             </s:param>
         </s:url>
         <tbody>
         <tr>
-            <td><s:property value="#user.id"/></td>
-            <td><s:property value="#user.userName"/></td>
-            <td><s:property value="#user.password"/></td>
-            <td><s:property value="#user.enabled"/></td>
+            <td><s:property value="#user.user_id"/></td>
+            <td><s:property value="#user.user_password"/></td>
+            <td><s:property value="#user.home_directory"/></td>
+            <td><s:property value="#user.enableflag"/></td>
+            <td><s:property value="#user.write_permission"/></td>
+            <td><s:property value="#user.idle_time"/></td>
+            <td><s:property value="#user.upload_rate"/></td>
+            <td><s:property value="#user.download_rate"/></td>
+            <td><s:property value="#user.max_login_number"/></td>
+            <td><s:property value="#user.max_login_per_ip"/></td>
+            <td><s:property value="#user.email"/></td>
             <td><s:property value="#user.user_roles.authority"/></td>
             <td>
                 <s:a href="%{userEditURL}">
@@ -52,7 +66,7 @@ add users in table and some additional information about users actions<br>
                             <span aria-hidden="true">7</span>
                     </span>
                 </s:a>
-                <s:a href="%{userDeleteURL}">
+                <s:a href="%{userDeleteURL}" onclick="return confirmDelete">
                     <span class="icon gray small" data-icon="m" title="Delete" style="display: inline-block; ">
                             <span aria-hidden="true">m</span>
                     </span>
